Complete the missing synopsis
Now it is possible to use cpd/cpc in the typical cp manner.
No longer the need to give the target as first argument, can
rename when copy one file and so on

Add the possibility to give -t as target directory. I'm not sure
why GNU cp works this way, I think it would work without the need
for -t too